Move book credit component
